[chrome] #319 Chrome plugin for KodeBeagle.
This should also set the basis for building snippet paste and search in web ui.

Implemented a chrome extension using chrome API. This extensions helps user to search KodeBeagle for any related code samples. Make a selection, right click and an option appears to search with KodeBeagle.

Code enhancements to filter java lang package elements before constructing the KodeBeagle search URL. In order to implement this we have introduced a java_lang_pkg.json file which holds all the necessary elements which needs to be filtered.