[core] #437 Updated the spark job to process repos with scala files too.
In current state the code for parsing and indexing scala files is very inefficient.
This choice was made because high rated repository for scala are very few and
investing time to make parsing less time consuming felt like unjustified effort.
However, I am still open to improvements and suggestion in terms of code contributions.

Scala parser is less accurate too, this can be upgraded by using scaliriform type of parser.
But unless we compile a scala file, it is not possible to have type information for everything.